Add manifest for istio-cni-windows
