+++ cni/pkg/iptables/iptables_e2e_linux_test.go
@@ -1,3 +1,7 @@
+//
+//go:build linux
+// +build linux
+
 // Copyright Istio Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
