+++ cni/pkg/iptables/iptables_unspecified.go
@@ -1,5 +1,5 @@
-//go:build !linux
-// +build !linux
+//go:build !linux && !windows
+// +build !linux,!windows
 
 // Copyright Istio Authors
 //
