+++ cni/pkg/iptables/test/Dockerfile
@@ -1,0 +1,5 @@
+FROM mcr.microsoft.com/windows/servercore:ltsc2022
+
+COPY ./hcn_test.exe hcn_test.exe
+
+ENTRYPOINT ["hcn_test.exe"]
