+++ cni/pkg/nodeagent/net.go
@@ -1,6 +1,6 @@
 // Copyright Istio Authors
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
+// Licensed under the Apache https://github.com/keithmattix/istio-cni-windowsLicense, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
