+++ cni/pkg/nodeagent/podcgroupns_linux_test.go
@@ -49,7 +49,7 @@ func TestWithProcFs(t *testing.T) {
 	}
 
 	expectedUID := "863b91d4-4b68-4efa-917f-4b560e3e86aa"
-	if podUIDNetns[expectedUID] == (WorkloadInfo{}) {
+	if podUIDNetns[expectedUID] == (workloadInfo{}) {
 		t.Fatal("expected to find pod netns under pod uid")
 	}
